API changes has been done

+++ src/test/java/pojo/Location.java
@@ -1,19 +1,19 @@
 package pojo;
 
 public class Location {
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
